requirements.txt
streamlit
faiss-cpu
sentence-transformers
pandas
random
pickle
